+++ SerbleChat.Backend/Database/Repos/IGuildRepo.cs
@@ -4,35 +4,35 @@ using SerbleChat.Backend.Schemas;
 namespace SerbleChat.Backend.Database.Repos;
 
 public interface IGuildRepo {
-    Task<Guild?> GetGuild(int id);
+    Task<Guild?> GetGuild(long id);
     Task<Guild[]> GetGuildsForUser(string userId);
     Task CreateGuild(Guild guild);
-    Task DeleteGuild(int id);
+    Task DeleteGuild(long id);
     Task UpdateGuild(Guild guild);
-    Task AddGuildMember(int guildId, string userId);
-    Task<bool> IsGuildMember(int guildId, string userId);
-    Task<ChatUser[]> GetGuildMembers(int guildId);
+    Task AddGuildMember(long guildId, string userId);
+    Task<bool> IsGuildMember(long guildId, string userId);
+    Task<ChatUser[]> GetGuildMembers(long guildId);
     
-    Task<GuildChannel?> GetGuildChannel(int channelId);
+    Task<GuildChannel?> GetGuildChannel(long channelId);
     Task CreateGuildChannel(GuildChannel channel);
-    Task DeleteGuildChannel(int id);
+    Task DeleteGuildChannel(long id);
     Task UpdateGuildChannel(GuildChannel channel);
-    Task<GuildChannel[]> GetGuildChannels(int guildId);
-    Task<Channel[]> GetGuildChannelsAsChannels(int guildId);
-    Task<ChatUser[]> GetGuildChannelMembers(int channelId);
-    Task<GuildMemberResponse[]> GetGuildChannelMembersDetails(int channelId);
+    Task<GuildChannel[]> GetGuildChannels(long guildId);
+    Task<Channel[]> GetGuildChannelsAsChannels(long guildId);
+    Task<ChatUser[]> GetGuildChannelMembers(long channelId);
+    Task<GuildMemberResponse[]> GetGuildChannelMembersDetails(long channelId);
     
-    Task<ChannelPermissionOverride[]> GetChannelPermissionOverrides(int channelId);
-    Task<ChannelPermissionOverride?> GetChannelPermissionOverride(int id);
+    Task<ChannelPermissionOverride[]> GetChannelPermissionOverrides(long channelId);
+    Task<ChannelPermissionOverride?> GetChannelPermissionOverride(long id);
     Task CreateChannelPermissionOverride(ChannelPermissionOverride permissionOverride);
-    Task DeleteChannelPermissionOverride(int id);
+    Task DeleteChannelPermissionOverride(long id);
     Task UpdateChannelPermissionOverride(ChannelPermissionOverride permissionOverride);
     
     Task CreateInvite(GuildInvite invite);
-    Task<GuildInvite?> GetInvite(int id);
-    Task DeleteInvite(int id);
-    Task<GuildInvite[]> GetGuildInvites(int guildId);
+    Task<GuildInvite?> GetInvite(long id);
+    Task DeleteInvite(long id);
+    Task<GuildInvite[]> GetGuildInvites(long guildId);
     
-    public Task<GuildPermissions> GetUserPermissions(string userId, int guildId, int channelId = -1);
-    public Task<Dictionary<int, GuildPermissions>> GetUserPermissionsForGuild(string userId, int guildId);
+    public Task<GuildPermissions> GetUserPermissions(string userId, long guildId, long channelId = -1);
+    public Task<Dictionary<long, GuildPermissions>> GetUserPermissionsForGuild(string userId, long guildId);
 }
