+++ SerbleChat.Backend/Database/Repos/IMessageRepo.cs
@@ -3,9 +3,9 @@ using SerbleChat.Backend.Database.Structs;
 namespace SerbleChat.Backend.Database.Repos;
 
 public interface IMessageRepo {
-    Task<List<Message>> GetMessages(int channelId, int limit = 50, int offset = 0);
-    Task<Message?> GetMessage(int id);
+    Task<List<Message>> GetMessages(long channelId, int limit = 50, int offset = 0);
+    Task<Message?> GetMessage(long id);
     Task CreateMessage(Message message);
     Task UpdateMessage(Message message);
-    Task DeleteMessage(int id);
+    Task DeleteMessage(long id);
 }
