+++ SerbleChat.Backend/Database/Repos/IRoleRepo.cs
@@ -4,13 +4,13 @@ using SerbleChat.Backend.Schemas;
 namespace SerbleChat.Backend.Database.Repos;
 
 public interface IRoleRepo {
-    public Task<Role?> GetRole(int id);
-    public Task<Role[]> GetGuildRoles(int guildId);
+    public Task<Role?> GetRole(long id);
+    public Task<Role[]> GetGuildRoles(long guildId);
     public Task CreateRole(Role role);
     public Task UpdateRole(Role role);
-    public Task DeleteRole(int id);
+    public Task DeleteRole(long id);
     
-    public Task<Role[]> GetUserRolesInGuild(string userId, int guildId);
-    public Task AddUserRole(int roleId, string userId);
-    public Task RemoveUserRole(int roleId, string userId);
+    public Task<Role[]> GetUserRolesInGuild(string userId, long guildId);
+    public Task AddUserRole(long roleId, string userId);
+    public Task RemoveUserRole(long roleId, string userId);
 }
