+++ SerbleChat.Backend/Database/Repos/IUserRepo.cs
@@ -18,20 +18,20 @@ public interface IUserRepo {
     Task<string> GetClientOptions(string userId);
     Task SetClientOptions(string userId, string options);
     
-    Task<UserChannelNotificationPreferences> GetChannelNotificationPreferences(string userId, int channelId);
-    Task<Dictionary<int, UserChannelNotificationPreferences>> GetAllChannelNotificationPreferences(string userId);
-    Task<UserGuildNotificationPreferences> GetUserGuildNotificationPreferences(string userId, int guildId);
-    Task<Dictionary<int, UserGuildNotificationPreferences>> GetAllUserGuildNotificationPreferences(string userId);
-    Task SetChannelNotificationPreferences(string userId, int channelId, UserChannelNotificationPreferences preferences);
-    Task SetUserGuildNotificationPreferences(string userId, int guildId, UserGuildNotificationPreferences preferences);
+    Task<UserChannelNotificationPreferences> GetChannelNotificationPreferences(string userId, long channelId);
+    Task<Dictionary<long, UserChannelNotificationPreferences>> GetAllChannelNotificationPreferences(string userId);
+    Task<UserGuildNotificationPreferences> GetUserGuildNotificationPreferences(string userId, long guildId);
+    Task<Dictionary<long, UserGuildNotificationPreferences>> GetAllUserGuildNotificationPreferences(string userId);
+    Task SetChannelNotificationPreferences(string userId, long channelId, UserChannelNotificationPreferences preferences);
+    Task SetUserGuildNotificationPreferences(string userId, long guildId, UserGuildNotificationPreferences preferences);
     
     // mass notification prefs gets
-    Task<Dictionary<string, UserChannelNotificationPreferences>> GetUsersChannelNotificationPreferences(IEnumerable<string> userIds, int channelId);
-    Task<Dictionary<string, UserGuildNotificationPreferences>> GetUsersGuildNotificationPreferences(IEnumerable<string> userIds, int guildId);
+    Task<Dictionary<string, UserChannelNotificationPreferences>> GetUsersChannelNotificationPreferences(IEnumerable<string> userIds, long channelId);
+    Task<Dictionary<string, UserGuildNotificationPreferences>> GetUsersGuildNotificationPreferences(IEnumerable<string> userIds, long guildId);
     
     Task CreateWebNotificationSubscription(UserWebNotificationHook subscription);
     Task<IEnumerable<UserWebNotificationHook>> GetWebNotificationSubscriptions(string userId);
-    Task DeleteWebNotificationSubscriptions(params int[] subscriptionId);
+    Task DeleteWebNotificationSubscriptions(params long[] subscriptionId);
     Task<Dictionary<string, IEnumerable<UserWebNotificationHook>>> GetUsersWebNotificationSubscriptions(IEnumerable<string> userIds);
     
     Task<PublicUserResponse> CompilePublicUserResponse(ChatUser user);
