+++ SerbleChat.Backend/Database/Repos/Impl/MessageRepo.cs
@@ -5,7 +5,7 @@ namespace SerbleChat.Backend.Database.Repos.Impl;
 
 public class MessageRepo(ChatDatabaseContext context) : IMessageRepo {
     
-    public async Task<List<Message>> GetMessages(int channelId, int limit = 50, int offset = 0) {
+    public async Task<List<Message>> GetMessages(long channelId, int limit = 50, int offset = 0) {
         return await context.Messages
             .Where(m => m.ChannelId == channelId)
             .OrderByDescending(m => m.CreatedAt)
@@ -14,7 +14,7 @@ public class MessageRepo(ChatDatabaseContext context) : IMessageRepo {
             .ToListAsync();
     }
 
-    public async Task<Message?> GetMessage(int id) {
+    public async Task<Message?> GetMessage(long id) {
         return await context.Messages.FindAsync(id);
     }
 
@@ -28,7 +28,7 @@ public class MessageRepo(ChatDatabaseContext context) : IMessageRepo {
         return context.SaveChangesAsync();
     }
 
-    public Task DeleteMessage(int id) {
+    public Task DeleteMessage(long id) {
         return context.Messages.Where(m => m.Id == id).ExecuteDeleteAsync();
     }
 }
