+++ SerbleChat.Backend/Database/Structs/MessageMention.cs
@@ -5,10 +5,10 @@ namespace SerbleChat.Backend.Database.Structs;
 
 public class MessageMention {
     [Key]
-    public int Id { get; set; }
+    public long Id { get; set; }
     
     [ForeignKey(nameof(MessageNavigation))]
-    public int MessageId { get; set; }
+    public long MessageId { get; set; }
     
     [StringLength(64), ForeignKey(nameof(UserNavigation))]
     public string UserId { get; set; } = null!;
